streaming_generator/requirements.txt
click==7.1.2
fastapi==0.65.1
h11==0.12.0
httptools==0.1.2
Jinja2==3.0.0
kafka-python==2.0.2
MarkupSafe==2.0.0
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
retrying==1.3.3
six==1.16.0
starlette==0.14.2
swagger-ui-py==0.3.0
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
